PAPP-27485: Microsoft Azure Compute: Feature - Adding dynamic subscription ID to actions (#5)
* Added functionality for dynamic Subscription IDs

* Updated readme

* Enhance code as per devloper best practices

* Add release notes

* Update error traceback flow

* Remove unwanted logs for state file

* Remove the unnecessary check for a corrupt state file and update the action description link for the run command

* Handling corrupt state file

* Update min_phantom_version and remove unwanted log statement

* Update Django wheel version for resolving fossa report

* Resolving semgrep issue

* Resolving semgrep failure

* Update release note and description of subscription_id action parameter

* Update error_code for while sanitize exception message

* Rename file names as per repo name

* fixing prodsec-scans pipeline

* Rename logo files

* Update view files name in json
