+++ .pre-commit-config.yaml
@@ -5,7 +5,7 @@ repos:
     -   id: org-hook
     -   id: package-app-dependencies
 -   repo: https://github.com/Yelp/detect-secrets
-    rev: v1.2.0
+    rev: v1.3.0
     hooks:
     -   id: detect-secrets
         args: ['--no-verify', '--exclude-files', '^microsoftazurevmmanagement.json$']
