+++ microsoftazurecompute_list_virtual_networks.html
@@ -1,0 +1,210 @@
+{% extends 'widgets/widget_template.html' %}
+{% load custom_template %}
+
+{% block custom_title_prop %}{% if title_logo %}style="background-size: auto 60%; background-position: 50%;
+background-repeat: no-repeat; background-image: url('/app_resource/{{ title_logo }}');"{% endif %}{% endblock %}
+{% block title1 %}{{ title1 }}{% endblock %}
+{% block title2 %}{{ title2 }}{% endblock %}
+{% block custom_tools %}
+{% endblock %}
+
+{% block widget_content %}
+<!-- Main Start Block -->
+
+<!-- File: microsoftazurecompute_list_virtual_networks.html
+  Copyright (c) 2019-2022 Splunk Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software distributed under
+the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+either express or implied. See the License for the specific language governing permissions
+and limitations under the License.
+-->
+
+<style>
+    .microsoftazurevmmanagement-app a:hover {
+        text-decoration: underline;
+    }
+
+    .microsoftazurevmmanagement-app .wf-table-vertical {
+        width: initial;
+        font-size: 12px;
+    }
+
+    .microsoftazurevmmanagement-app .wf-table-vertical td {
+        padding: 5px;
+        border: 1px solid;
+    }
+
+    .microsoftazurevmmanagement-app .wf-table-horizontal {
+        margin-right: 10px;
+        width: initial;
+        border: 1px solid;
+        font-size: 12px;
+    }
+
+    .microsoftazurevmmanagement-app .wf-table-horizontal th {
+        text-align: center;
+        border: 1px solid;
+        text-transform: uppercase;
+        font-weight: normal;
+        padding: 5px;
+    }
+
+    .microsoftazurevmmanagement-app .wf-table-horizontal td {
+        border: 1px solid;
+        padding: 5px;
+        padding-left: 4px;
+    }
+
+    .microsoftazurevmmanagement-app .wf-h3-style {
+        font-size: 20px
+    }
+
+    .microsoftazurevmmanagement-app .wf-h4-style {
+        font-size: 16px
+    }
+
+    .microsoftazurevmmanagement-app .wf-h5-style {
+        font-size: 14px
+    }
+
+    .microsoftazurevmmanagement-app .wf-subheader-style {
+        font-size: 12px
+    }
+</style>
+<div class="microsoftazurevmmanagement-app" style="overflow: auto; width: 100%; height: 100%; padding-left:10px; padding-right:10px">
+    <!-- Main Div -->
+    {% for result in results %}
+        <!-- loop for each result -->
+        {% if result.data %}
+            <div class="info">
+            <h4 class="wf-h4-style">Info</h4>
+            <table class="wf-table-vertical">
+                {% if result.param.resource_group_name %}
+                <tr>
+                    <td>Resource Group Name</td>
+                    <td>
+                        <a href="javascript:;" onclick="context_menu(this, [{'contains': ['vm management resource group'],
+                        'value':'{{ result.param.resource_group_name }}' }], 0,
+                        {{ container.id }}, null, false);">
+                            {{ result.param.resource_group_name }}
+                            &nbsp;
+                            <span class="fa fa-caret-down" style="font-size: smaller;"></span>
+                        </a>
+                    </td>
+                </tr>
+                {% endif %}
+            </table>
+            </div>
+            <h3 class="wf-h3-style">Virtual Networks</h3>
+            {% for data in result.data %}
+                <!-- Tags -->
+                <h3 class="wf-h4-style">{{ data.name }}</h3>
+                <table class="wf-table-vertical">
+                    <tr>
+                        <th>Network Name</th>
+                        <td>
+                            <a href="javascript:;" onclick="context_menu(this, [{'contains': ['vm management virtual network'],
+                            'value':'{{ data.name }}' }], 0, {{ container.id }}, null, false);">
+                            {{ data.name }}
+                            &nbsp;
+                            <span class="fa fa-caret-down" style="font-size: smaller;"></span>
+                            </a>
+                        </td>
+                    </tr>
+                    <tr>
+                        <th>Etag</th>
+                        <td>{{ data.etag }}</td>
+                    </tr>
+                    <tr>
+                        <th>Location</th>
+                        <td>{{ data.location }}</td>
+                    </tr>
+                    <tr>
+                        <th>Type</th>
+                        <td>{{ data.type }}</td>
+                    </tr>
+                    <tr>
+                        <th>ID</th>
+                        <td>{{ data.id }}</td>
+                    </tr>
+                    <tr>
+                        <th>Enable VM Protection</th>
+                        <td>{{ data.properties.enableVmProtection }}</td>
+                    </tr>
+                    <tr>
+                        <th>Enable DDOS Protection</th>
+                        <td>{{ data.properties.enableDdosProtection }}</td>
+                    </tr>
+                    <tr>
+                        <th>Resource GUID</th>
+                        <td>{{ data.properties.resourceGuid }}</td>
+                    </tr>
+                    <tr>
+                        <th>Provisioning State</th>
+                        <td>{{ data.properties.provisioningState }}</td>
+                    </tr>
+                </table>
+
+                {% if data.tags %}
+                    <br>
+                    <table class="wf-table-horizontal">
+                        <tr>
+                            <th>Tag Name</th>
+                            <th>Tag Value</th>
+                        </tr>
+                        {% for key, val in data.tags.items %}
+                        <tr>
+                            <td>{{ key }}</td>
+                            <td>{{ val }}</td>
+                        </tr>
+                        {% endfor %}
+                    </table>
+                {% endif %}
+
+                {% if data.properties.subnets %}
+                    <br>
+                    {% for val in data.properties.subnets %}
+                        <table class="wf-table-vertical">
+                            <tr>
+                                <th>Subnet Name</th>
+                                <td>{{ val.name }}</td>
+                            </tr>
+                            <tr>
+                                <th>Etag</th>
+                                <td>{{ val.etag }}</td>
+                            </tr>
+                            <tr>
+                                <th>Type</th>
+                                <td>{{ val.type }}</td>
+                            </tr>
+                            <tr>
+                                <th>ID</th>
+                                <td>{{ val.id }}</td>
+                            </tr>
+                        </table>
+                    {% endfor %}
+
+                {% endif %}
+
+                <br>
+
+            {% endfor %}
+            <!-- loop for each data end -->
+        {% else %}
+            <h4 class="wf-h4-style">No data found</h4>
+        {% endif %}
+        <!-- loop for each data end -->
+
+    {% endfor %}
+    <!-- loop for each result end -->
+</div> <!-- Main Div -->
+
+{% endblock %}
+<!-- Main Start Block -->
