--- microsoftazurevmmanagement_view.py
@@ -1,200 +1,0 @@
-# File: microsoftazurevmmanagement_view.py
-#
-# Copyright (c) 2019-2022 Splunk Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software distributed under
-# the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
-# either express or implied. See the License for the specific language governing permissions
-# and limitations under the License.
-def _get_ctx_result(provides, result):
-    """ Function that parse data.
-
-    :param provides: action name
-    :param result: result
-    :return: context response
-    """
-
-    ctx_result = {}
-
-    param = result.get_param()
-    summary = result.get_summary()
-    data = result.get_data()
-
-    ctx_result['param'] = param
-    if summary:
-        ctx_result['summary'] = summary
-
-    if not data:
-        ctx_result['data'] = {}
-        return ctx_result
-
-    ctx_result['action'] = provides
-    ctx_result['data'] = data
-
-    return ctx_result
-
-
-def display_list_tags(provides, all_app_runs, context):
-    """ Function that display flows.
-
-    :param provides: action name
-    :param all_app_runs: all_app_runs
-    :param context: context
-    :return: html page name
-    """
-
-    context['results'] = results = []
-    for summary, action_results in all_app_runs:
-        for result in action_results:
-            ctx_result = _get_ctx_result(provides, result)
-            if not ctx_result:
-                continue
-            results.append(ctx_result)
-
-    return "microsoftazurevmmanagement_list_tags.html"
-
-
-def display_list_virtual_networks(provides, all_app_runs, context):
-    """ Function that display flows.
-
-    :param provides: action name
-    :param all_app_runs: all_app_runs
-    :param context: context
-    :return: html page name
-    """
-
-    context['results'] = results = []
-    for summary, action_results in all_app_runs:
-        for result in action_results:
-            ctx_result = _get_ctx_result(provides, result)
-            if not ctx_result:
-                continue
-            results.append(ctx_result)
-
-    return "microsoftazurevmmanagement_list_virtual_networks.html"
-
-
-def display_snapshot_vm(provides, all_app_runs, context):
-    """ Function that display flows.
-
-    :param provides: action name
-    :param all_app_runs: all_app_runs
-    :param context: context
-    :return: html page name
-    """
-
-    context['results'] = results = []
-    for summary, action_results in all_app_runs:
-        for result in action_results:
-            ctx_result = _get_ctx_result(provides, result)
-            if not ctx_result:
-                continue
-            results.append(ctx_result)
-
-    return "microsoftazurevmmanagement_snapshot_vm.html"
-
-
-def display_add_application_security_group(provides, all_app_runs, context):
-    """ Function that display flows.
-
-    :param provides: action name
-    :param all_app_runs: all_app_runs
-    :param context: context
-    :return: html page name
-    """
-
-    context['results'] = results = []
-    for summary, action_results in all_app_runs:
-        for result in action_results:
-            ctx_result = _get_ctx_result(provides, result)
-            if not ctx_result:
-                continue
-            results.append(ctx_result)
-
-    return "microsoftazurevmmanagement_add_application_security_group.html"
-
-
-def display_add_network_security_group(provides, all_app_runs, context):
-    """ Function that display flows.
-
-    :param provides: action name
-    :param all_app_runs: all_app_runs
-    :param context: context
-    :return: html page name
-    """
-
-    context['results'] = results = []
-    for summary, action_results in all_app_runs:
-        for result in action_results:
-            ctx_result = _get_ctx_result(provides, result)
-            if not ctx_result:
-                continue
-            results.append(ctx_result)
-
-    return "microsoftazurevmmanagement_add_network_security_group.html"
-
-
-def display_list_vms(provides, all_app_runs, context):
-    """ Function that display flows.
-
-    :param provides: action name
-    :param all_app_runs: all_app_runs
-    :param context: context
-    :return: html page name
-    """
-
-    context['results'] = results = []
-    for summary, action_results in all_app_runs:
-        for result in action_results:
-            ctx_result = _get_ctx_result(provides, result)
-            if not ctx_result:
-                continue
-            results.append(ctx_result)
-
-    return "microsoftazurevmmanagement_list_vms.html"
-
-
-def display_list_snapshots(provides, all_app_runs, context):
-    """ Function that display flows.
-
-    :param provides: action name
-    :param all_app_runs: all_app_runs
-    :param context: context
-    :return: html page name
-    """
-
-    context['results'] = results = []
-    for summary, action_results in all_app_runs:
-        for result in action_results:
-            ctx_result = _get_ctx_result(provides, result)
-            if not ctx_result:
-                continue
-            results.append(ctx_result)
-
-    return "microsoftazurevmmanagement_list_snapshots.html"
-
-
-def display_get_system_info(provides, all_app_runs, context):
-    """ Function that display flows.
-
-    :param provides: action name
-    :param all_app_runs: all_app_runs
-    :param context: context
-    :return: html page name
-    """
-
-    context['results'] = results = []
-    for summary, action_results in all_app_runs:
-        for result in action_results:
-            ctx_result = _get_ctx_result(provides, result)
-            if not ctx_result:
-                continue
-            results.append(ctx_result)
-
-    return "microsoftazurevmmanagement_get_system_info.html"
