+++ release_notes/unreleased.md
@@ -1,1 +1,3 @@
 **Unreleased**
+* Added action level optional parameter 'subscription ID' to override configuration 'subscription ID'
+* Added encryption for the sensitive values stored in the state file
