+++ requirements.txt
@@ -1,3 +1,3 @@
 beautifulsoup4==4.9.1
-Django==3.2.13
+Django==3.2.16
 requests==2.25.0
